Reduce rounding errors in l2_sqr_float()
This change improves the floating point accuracy of the l2_sqr operation
on float32 values. With divide and conquer summation, we reduce roundoff
errors to logarithmic growth. For example, on an array with 32768 floats
of random values on the interval (-1,1), the accuracy of naïve summation
degrades to 1e-3 whereas this algorithm maintains 1e-7 accuracy compared
to an implementation that uses a float 64 compute type. Divide / conquer
also enables better vectorization.

    1.10731 ns 100x l2_sqr_float_avx
    1.26103 ns 100x l2_sqr_float in ANSI C (NEW implementation)
    1.65579 ns 100x l2_sqr_float in ANSI C (old implementation)
    1.87289 ns 100x l2_sqr_double

The benchmarks above, measured in nanoseconds per float show we can gain
nearly all the performance benefits of intrinsics without the complexity
therefore this change removes the AVX and NEON implementations, since it
isn't worth compromising standards conformance and clarity for just 10%.